# rpc/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: Tendermint RPC
  contact:
    name: Tendermint RPC
    url: https://github.com/tendermint/tendermint/issues/new/choose
  description: |
    Tendermint supports the following RPC protocols:

    * URI over HTTP
    * JSONRPC over HTTP
    * JSONRPC over websockets

    ## Configuration

    RPC can be configured by tuning parameters under `[rpc]` table in the
    `$TMHOME/config/config.toml` file or by using the `--rpc.X` command-line
    flags.

    Default rpc listen address is `tcp://0.0.0.0:26657`.
    To set another address, set the `laddr` config parameter to desired value.
    CORS (Cross-Origin Resource Sharing) can be enabled by setting
    `cors_allowed_origins`, `cors_allowed_methods`, `cors_allowed_headers`
    config parameters.

    ## Arguments

    Arguments which expect strings or byte arrays may be passed as quoted
    strings, like `"abc"` or as `0x`-prefixed strings, like `0x616263`.

    ## URI/HTTP

    A REST like interface.

        curl localhost:26657/block?height=5

    ## JSONRPC/HTTP

    JSONRPC requests can be POST'd to the root RPC endpoint via HTTP.

        curl --header "Content-Type: application/json" --request POST --data '{"method": "block", "params": ["5"], "id": 1}' localhost:26657

    ## JSONRPC/websockets

    JSONRPC requests can be also made via websocket.
    The websocket endpoint is at `/websocket`, e.g. `localhost:26657/websocket`.
...

...

...
  license:
...
servers:
...
paths:
...
  /broadcast_tx_async:
    get:
      summary: Returns right away, with no response. Does not wait for CheckTx nor DeliverTx results.
      tags:
        - Tx
      operationId: broadcast_tx_async
      description: |
        If you want to be sure that the transaction is included in a block, you can
        subscribe for the result using JSONRPC via a websocket. See
        https://docs.tendermint.com/v0.34/app-dev/subscribing-to-events-via-websocket.html
        If you haven't received anything after a couple of blocks, resend it. If the
        same happens again, send it to some other node. A few reasons why it could
        happen:

        1. malicious node can drop or pretend it had committed your tx
        2. malicious proposer (not necessary the one you're communicating with) can
        drop transactions, which might become valid in the future
        (https://github.com/tendermint/tendermint/issues/3322)
        3. node can be offline

        Please refer to
        https://docs.tendermint.com/v0.34/tendermint-core/using-tendermint.html#formatting
        for formatting/encoding rules.
      parameters:
        - in: query
          name: tx
          required: true
          schema:
            type: string
            example: "123"
          description: The transaction
      responses:
        "200":
          description: empty answer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BroadcastTxResponse"
        "500":
          description: empty error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /broadcast_tx_commit:
    get:
      summary: Returns with the responses from CheckTx and DeliverTx.
      tags:
        - Tx
      operationId: broadcast_tx_commit
      description: |
        IMPORTANT: use only for testing and development. In production, use
        BroadcastTxSync or BroadcastTxAsync. You can subscribe for the transaction
        result using JSONRPC via a websocket. See
        https://docs.tendermint.com/v0.34/app-dev/subscribing-to-events-via-websocket.html

        CONTRACT: only returns error if mempool.CheckTx() errs or if we timeout
        waiting for tx to commit.

        If CheckTx or DeliverTx fail, no error will be returned, but the returned result
        will contain a non-OK ABCI code.

        Please refer to
        https://docs.tendermint.com/v0.34/tendermint-core/using-tendermint.html#formatting
        for formatting/encoding rules.
      parameters:
        - in: query
          name: tx
          required: true
          schema:
            type: string
            example: "785"
          description: The transaction
      responses:
        "200":
          description: empty answer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BroadcastTxCommitResponse"
        "500":
          description: empty error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /tx:
    get:
      summary: Get transactions by hash
      operationId: tx
      parameters:
        - in: query
          name: hash
          description: hash of transaction to retrieve
          required: true
          schema:
            type: string
            example: "0xD70952032620CC4E2737EB8AC379806359D8E0B17B0488F627997A0B043ABDED"
        - in: query
          name: prove
          description: Include proofs of the transaction's inclusion in the block
          required: false
          schema:
            type: boolean
            example: true
            default: false
      tags:
        - Info
      description: |
        Get a transaction

        Upon success, the `Cache-Control` header will be set with the default
        maximum age.
      responses:
        "200":
          description: Get a transaction`
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TxResponse"
        "500":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /abci_query:
    get:
      summary: Query the application for some information.
      operationId: abci_query
      parameters:
        - in: query
          name: path
          description: Path to the data ("/a/b/c")
          required: true
          schema:
            type: string
            example: "/a/b/c"
        - in: query
          name: data
          description: Data
          required: true
          schema:
            type: string
            example: "IHAVENOIDEA"
        - in: query
          name: height
          description: Height (0 means latest)
          required: false
          schema:
            type: integer
            example: 1
            default: 0
        - in: query
          name: prove
          description: Include proofs of the transactions inclusion in the block
          required: false
          schema:
            type: boolean
            example: true
            default: false
      tags:
        - ABCI
      description: |
        Query the application for some information.
      responses:
        "200":
          description: Response of the submitted query
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ABCIQueryResponse"
        "500":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
    JSONRPC:
      type: object
      properties:
        id:
          type: integer
          example: 0
        jsonrpc:
          type: string
          example: "2.0"
...
    ErrorResponse:
      description: Error Response
      allOf:
        - $ref: "#/components/schemas/JSONRPC"
        - type: object
          properties:
            error:
              type: string
              example: "Description of failure"
...
    TxResponse:
      type: object
      required:
        - "jsonrpc"
        - "id"
        - "result"
      properties:
        jsonrpc:
          type: string
          example: "2.0"
        id:
          type: integer
          example: 0
        result:
          required:
            - "hash"
            - "height"
            - "index"
            - "tx_result"
            - "tx"
          properties:
            hash:
              type: string
              example: "D70952032620CC4E2737EB8AC379806359D8E0B17B0488F627997A0B043ABDED"
            height:
              type: string
              example: "1000"
            index:
              type: integer
              example: 0
            tx_result:
              required:
                - "log"
                - "gas_wanted"
                - "gas_used"
                - "tags"
              properties:
                log:
                  type: string
                  example: '[{"msg_index":"0","success":true,"log":""}]'
                gas_wanted:
                  type: string
                  example: "200000"
                gas_used:
                  type: string
                  example: "28596"
                tags:
                  type: array
                  items:
                    $ref: "#/components/schemas/Event"
              type: object
            tx:
              type: string
              example: "5wHwYl3uCkaoo2GaChQmSIu8hxpJxLcCuIi8fiHN4TMwrRIU/Af1cEG7Rcs/6LjTl7YjRSymJfYaFAoFdWF0b20SCzE0OTk5OTk1MDAwEhMKDQoFdWF0b20SBDUwMDAQwJoMGmoKJuta6YchAwswBShaB1wkZBctLIhYqBC3JrAI28XGzxP+rVEticGEEkAc+khTkKL9CDE47aDvjEHvUNt+izJfT4KVF2v2JkC+bmlH9K08q3PqHeMI9Z5up+XMusnTqlP985KF+SI5J3ZOIhhNYWRlIGJ5IENpcmNsZSB3aXRoIGxvdmU="
          type: object
...
    ABCIQueryResponse:
      type: object
      required:
        - "error"
        - "result"
        - "id"
        - "jsonrpc"
      properties:
        error:
          type: string
          example: ""
        result:
          required:
            - "response"
          properties:
            response:
              required:
                - "log"
                - "height"
                - "proof"
                - "value"
                - "key"
                - "index"
                - "code"
              properties:
                log:
                  type: string
                  example: "exists"
                height:
                  type: string
                  example: "0"
                proof:
                  type: string
                  example: "010114FED0DAD959F36091AD761C922ABA3CBF1D8349990101020103011406AA2262E2F448242DF2C2607C3CDC705313EE3B0001149D16177BC71E445476174622EA559715C293740C"
                value:
                  type: string
                  example: "61626364"
                key:
                  type: string
                  example: "61626364"
                index:
                  type: string
                  example: "-1"
                code:
                  type: string
                  example: "0"
              type: object
          type: object
        id:
          type: integer
          example: 0
        jsonrpc:
          type: string
          example: "2.0"
...
    BroadcastTxCommitResponse:
      type: object
      required:
        - "error"
        - "result"
        - "id"
        - "jsonrpc"
      properties:
        error:
          type: string
          example: ""
        result:
          required:
            - "height"
            - "hash"
            - "deliver_tx"
            - "check_tx"
          properties:
            height:
              type: string
              example: "26682"
            hash:
              type: string
              example: "75CA0F856A4DA078FC4911580360E70CEFB2EBEE"
            deliver_tx:
              required:
                - "log"
                - "data"
                - "code"
              properties:
                log:
                  type: string
                  example: ""
                data:
                  type: string
                  example: ""
                code:
                  type: string
                  example: "0"
              type: object
            check_tx:
              required:
                - "log"
                - "data"
                - "code"
              properties:
                log:
                  type: string
                  example: ""
                data:
                  type: string
                  example: ""
                code:
                  type: string
                  example: "0"
              type: object
          type: object
        id:
          type: integer
          example: 0
        jsonrpc:
          type: string
          example: "2.0"
...
    BroadcastTxResponse:
      type: object
      required:
        - "jsonrpc"
        - "id"
        - "result"
        - "error"
      properties:
        jsonrpc:
          type: string
          example: "2.0"
        id:
          type: integer
          example: 0
        result:
          required:
            - "code"
            - "data"
            - "log"
            - "hash"
          properties:
            code:
              type: string
              example: "0"
            data:
              type: string
              example: ""
            log:
              type: string
              example: ""
            codespace:
              type: string
              example: "ibc"
            hash:
              type: string
              example: "0D33F2F03A5234F38706E43004489E061AC40A2E"
          type: object
        error:
          type: string
          example: ""
...
    Event:
      type: object
      properties:
        key:
          type: string
          example: "YWN0aW9u"
        value:
          type: string
          example: "c2VuZA=="
        index:
          type: boolean
          example: false

    # Block Header
...